Edit Redis and add MySQL to Troubleshooting.md
These parts have been edited based on my experiences. It is impossible for a server to have the same allocation port twice, this is an issue that blocks you from connecting servers to a proxy. So Pterodactyl by default allows containers to access your local network through 172.18.0.1 since 127.0.0.1 inside the container would be the container itself.